Added new stuff

+++ app/Http/Controllers/AdminController.php
@@ -8,6 +8,7 @@ use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Str;
 use App\Admin;
 use Auth;
+use Spatie\Permission\Models\Role;
 
 class AdminController extends Controller
 {
@@ -19,10 +20,10 @@ class AdminController extends Controller
     public function index()
     {
         $admins = Admin::where('school_id','=',Auth::user()->school_id)->paginate(20);
-        if(Auth::guard('admin'))
+        if(Auth::guard('admin')->user()->hasPermissionTo('view-admin', 'admin'))
             return view('admin.admin.index')->with('admins',$admins);
         else
-            return redirect('/home')->with('error','No access');
+            return redirect(route('admin.home'))->with('error',__('messages.noauthorization'));
     }
 
     /**
@@ -32,10 +33,11 @@ class AdminController extends Controller
      */
     public function create()
     {
-        if(Auth::guard('admin'))
-            return view('admin.admin.create');
+        $roles = Role::where('school_id','=',Auth::guard('admin')->user()->school_id)->get();
+        if(Auth::guard('admin')->user()->hasPermissionTo('create-admin', 'admin'))
+            return view('admin.admin.create')->with('roles',$roles);
         else
-            return redirect('/home')->with('error','No access');
+            return redirect(route('admin.home'))->with('error',__('messages.noauthorization'));
     }
 
     /**
@@ -46,48 +48,57 @@ class AdminController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request,[
-            'Emri'=> 'required|min:2|string',
-            'Emri_Prindit'=> 'required|min:2|string',
-            'Mbiemri'=> 'required|min:2|string',
-            'Adresa'=> 'required|min:2|string',
-            'Qyteti'=> 'required|min:2|string',
-            'Vendbanimi'=> 'required|min:2|string',
-            'Grada'=> 'required|min:2|string',
-            'Nr_Telefonit'=> 'required|min:9|numeric',
-            'Data_e_lindjes'=> 'required|date',
-            'email' => 'required|email|unique:admins',
-            'password' => 'required|string|min:6|confirmed',
-            'Foto' =>'image|nullable|max:1999',
-        ]);
-        if($request->hasFile('Foto'))
+        if(Auth::guard('admin')->user()->hasPermissionTo('create-admin', 'admin'))
         {
-            $fileNamewithExt = $request->file('Foto')->getClientOriginalName();
-            $fileName = pathInfo($fileNamewithExt, PATHINFO_FILENAME);
-            $extension = $request->file('Foto')->getClientOriginalExtension();
-            $fileNametoStore = 'foto-'.Str::random(25);
-            $request->file('Foto')->move(public_path('img'), $fileNametoStore);
+            $this->validate($request,[
+                'Emri'=> 'required|min:2|string',
+                'Emri_Prindit'=> 'required|min:2|string',
+                'Mbiemri'=> 'required|min:2|string',
+                'Adresa'=> 'required|min:2|string',
+                'Qyteti'=> 'required|min:2|string',
+                'Vendbanimi'=> 'required|min:2|string',
+                'Nr_Telefonit'=> 'required|min:9|numeric',
+                'Data_e_lindjes'=> 'required|date',
+                'email' => 'required|email|unique:admins',
+                'password' => 'required|string|min:6|confirmed',
+                'Foto' =>'image|nullable|max:1999',
+                'Roli'=> 'required',
+            ]);
+            if($request->hasFile('Foto'))
+            {
+                $fileNamewithExt = $request->file('Foto')->getClientOriginalName();
+                $fileName = pathInfo($fileNamewithExt, PATHINFO_FILENAME);
+                $extension = $request->file('Foto')->getClientOriginalExtension();
+                $fileNametoStore = 'foto-'.Str::random(25).'.'.$extension;
+                $request->file('Foto')->move(public_path('img'), $fileNametoStore);
+            }
+            else
+            {
+                $fileNametoStore = 'no-image';
+            }
+            $admin = new Admin;
+            $admin->fathers_name = $request->input('Emri_Prindit');
+            $admin->first_name = $request->input('Emri');
+            $admin->surname = $request->input('Mbiemri');
+            $admin->address = $request->input('Adresa');
+            $admin->birthday = $request->input('Data_e_lindjes');
+            $admin->city = $request->input('Qyteti');
+            $admin->residence = $request->input('Vendbanimi');
+            $admin->phone_nr = $request->input('Nr_Telefonit');
+            $admin->grade = $request->input('Roli');
+            $admin->gender = $request->input('Gjinia');
+            $admin->photo = $fileNametoStore;
+            $admin->school_id = Auth::guard('admin')->user()->school_id;
+            $admin->email = $request->input('email');
+            $admin->password = Hash::make($request->input('password'));
+            $admin->save();
+            $admin->assignRole($request->input('Roli'));
+            return redirect(route('admin.admin.index'))->with('success',__('messages.admin-add'));
         }
         else
         {
-            $fileNametoStore = 'no-image';
+            return redirect(route('admin.home'))->with('error',__('messages.noauthorization'));
         }
-        $admin = new Admin;
-        $admin->fathers_name = $request->input('Emri_Prindit');
-        $admin->first_name = $request->input('Emri');
-        $admin->surname = $request->input('Mbiemri');
-        $admin->address = $request->input('Adresa');
-        $admin->birthday = $request->input('Data_e_lindjes');
-        $admin->city = $request->input('Qyteti');
-        $admin->residence = $request->input('Vendbanimi');
-        $admin->phone_nr = $request->input('Nr_Telefonit');
-        $admin->grade = $request->input('Grada');;
-        $admin->photo = $fileNametoStore;
-        $admin->school_id = Auth::guard('admin')->user()->school_id;
-        $admin->email = $request->input('email');
-        $admin->password = Hash::make($request->input('password'));
-        $admin->save();
-        return redirect('/admin/admin')->with('success','U shtua Stafi');
     }
 
     /**
@@ -99,10 +110,10 @@ class AdminController extends Controller
     public function show($id)
     {
         $admin = Admin::findOrFail($id);
-        if(Auth::guard('admin'))
+        if(Auth::guard('admin')->user()->hasPermissionTo('view-admin', 'admin'))
             return view('admin.admin.show')->with('admin',$admin);
         else
-            return redirect('/home')->with('error','No access');
+            return redirect(route('admin.home'))->with('error',__('messages.noauthorization'));
     }
 
     /**
@@ -114,10 +125,11 @@ class AdminController extends Controller
     public function edit($id)
     {
         $admin = Admin::findOrFail($id);
-        if(Auth::guard('admin'))
-            return view('admin.admin.edit')->with('admin',$admin);
+        $roles = Role::where('school_id','=',Auth::guard('admin')->user()->school_id)->get();
+        if(Auth::guard('admin')->user()->hasPermissionTo('edit-admin', 'admin'))
+            return view('admin.admin.edit')->with('admin',$admin)->with('roles',$roles);
         else
-            return redirect('/home')->with('error','No access');
+            return redirect(route('admin.home'))->with('error',__('messages.noauthorization'));
     }
 
     /**
@@ -129,42 +141,50 @@ class AdminController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request,[
-            'Emri'=> 'required|min:2|string',
-            'Emri_Prindit'=> 'required|min:2|string',
-            'Mbiemri'=> 'required|min:2|string',
-            'Adresa'=> 'required|min:2|string',
-            'Qyteti'=> 'required|min:2|string',
-            'Vendbanimi'=> 'required|min:2|string',
-            'Nr_Telefonit'=> 'required|min:9|numeric',
-            'Data_e_lindjes'=> 'required|date',
-            'email' => 'required|email|unique:admins,email,'.$id,
-            'Foto' =>'image|nullable|max:1999',
-            'Grada'=> 'required|min:2|string',
-        ]);
-        $admin = Admin::find($id);
-        if($request->hasFile('Foto'))
+        if(Auth::guard('admin')->user()->hasPermissionTo('edit-admin', 'admin'))
         {
-            $fileNamewithExt = $request->file('Foto')->getClientOriginalName();
-            $fileName = pathInfo($fileNamewithExt, PATHINFO_FILENAME);
-            $extension = $request->file('Foto')->getClientOriginalExtension();
-            $fileNametoStore = 'foto-'.Str::random(25);
-            $request->file('Foto')->move(public_path('img'), $fileNametoStore);
-            $admin->photo = $fileNametoStore;
+            $this->validate($request,[
+                'Emri'=> 'required|min:2|string',
+                'Emri_Prindit'=> 'required|min:2|string',
+                'Mbiemri'=> 'required|min:2|string',
+                'Adresa'=> 'required|min:2|string',
+                'Qyteti'=> 'required|min:2|string',
+                'Vendbanimi'=> 'required|min:2|string',
+                'Nr_Telefonit'=> 'required|min:9|numeric',
+                'Data_e_lindjes'=> 'required|date',
+                'email' => 'required|email|unique:admins,email,'.$id,
+                'Foto' =>'image|nullable|max:1999',
+                'Roli'=> 'required',
+            ]);
+            $admin = Admin::find($id);
+            if($request->hasFile('Foto'))
+            {
+                $fileNamewithExt = $request->file('Foto')->getClientOriginalName();
+                $fileName = pathInfo($fileNamewithExt, PATHINFO_FILENAME);
+                $extension = $request->file('Foto')->getClientOriginalExtension();
+                $fileNametoStore = 'foto-'.Str::random(25).'.'.$extension;
+                $request->file('Foto')->move(public_path('img'), $fileNametoStore);
+                $admin->photo = $fileNametoStore;
+            }
+            $admin->first_name = $request->input('Emri');
+            $admin->fathers_name = $request->input('Emri_Prindit');
+            $admin->surname = $request->input('Mbiemri');
+            $admin->address = $request->input('Adresa');
+            $admin->birthday = $request->input('Data_e_lindjes');
+            $admin->city = $request->input('Qyteti');
+            $admin->residence = $request->input('Vendbanimi');
+            $admin->phone_nr = $request->input('Nr_Telefonit');
+            $admin->grade =  $request->input('Roli');
+            $admin->gender = $request->input('Gjinia');
+            $admin->school_id = Auth::guard('admin')->user()->school_id;
+            $admin->email = $request->input('email');
+            $admin->save();
+            $admin->syncRoles($request->input('Roli'));
+            return redirect(route('admin.admin.index'))->with('success',__('messages.admin-edit'));
         }
-        $admin->first_name = $request->input('Emri');
-        $admin->fathers_name = $request->input('Emri_Prindit');
-        $admin->surname = $request->input('Mbiemri');
-        $admin->address = $request->input('Adresa');
-        $admin->birthday = $request->input('Data_e_lindjes');
-        $admin->city = $request->input('Qyteti');
-        $admin->residence = $request->input('Vendbanimi');
-        $admin->phone_nr = $request->input('Nr_Telefonit');
-        $admin->grade =  $request->input('Grada');
-        $admin->school_id = Auth::guard('admin')->user()->school_id;
-        $admin->email = $request->input('email');
-        $admin->save();
-        return redirect('/admin/admin')->with('success','U ndryshua staffi');
+        else{
+            return redirect(route('admin.home'))->with('error',__('messages.noauthorization'));
+        }
     }
 
     /**
@@ -175,16 +195,21 @@ class AdminController extends Controller
      */
     public function destroy($id)
     {
-        $admin = Admin::find($id);
-        if($admin->grades()->get()->count() > 0)
-            $admin->grades()->delete();
-        if($admin->subjects()->get()->count() > 0)
-            $admin->subjects()->detach();
-        if($admin->schedules()->get()->count() > 0)
-            $admin->schedules()->delete();
-        if($admin->classroom()->get()->count() > 0)
-            $admin->classroom()->delete();
-        $admin->delete();
-        return redirect('/admin/admin')->with('success','U fshi stafi');
+        if(Auth::guard('admin')->user()->hasPermissionTo('delete-admin', 'admin'))
+        {
+            $admin = Admin::find($id);
+            if($admin->grades()->get()->count() > 0)
+                $admin->grades()->delete();
+            if($admin->subjects()->get()->count() > 0)
+                $admin->subjects()->detach();
+            if($admin->schedules()->get()->count() > 0)
+                $admin->schedules()->delete();
+            if($admin->classroom()->get()->count() > 0)
+                $admin->classroom()->delete();
+            $admin->delete();
+            return redirect(route('admin.admin.index'))->with('success',__('messages.admin-delete'));
+        }
+        else
+            return redirect(route('admin.home'))->with('error',__('messages.noauthorization'));
     }
 }
