+++ app/Http/Controllers/SubjectController.php
@@ -16,10 +16,10 @@ class SubjectController extends Controller
     public function index()
     {
         $subjects = Subject::where('school_id','=',Auth::user()->school_id)->paginate(20);
-        if(Auth::guard('admin'))
+        if(Auth::guard('admin')->user()->hasPermissionTo('view-subject', 'admin'))
             return view('admin.subject.index')->with('subjects',$subjects);
         else
-            return redirect('/home')->with('error','No access');
+            return redirect(route('admin.home'))->with('error',__('messages.noauthorization'));
     }
 
     /**
@@ -29,10 +29,10 @@ class SubjectController extends Controller
      */
     public function create()
     {
-        if(Auth::guard('admin'))
+        if(Auth::guard('admin')->user()->hasPermissionTo('create-subject', 'admin'))
             return view('admin.subject.create');
         else
-            return redirect('/home')->with('error','No access');
+            return redirect(route('admin.home'))->with('error',__('messages.noauthorization'));
     }
 
     /**
@@ -43,16 +43,20 @@ class SubjectController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request,[
-            'Viti'=> 'required|numeric',
-            'Emri'=> 'required|string|min:2',
-        ]);
-        $subject = new Subject;
-        $subject->year = $request->input('Viti');
-        $subject->name = $request->input('Emri');
-        $subject->school_id = Auth::guard('admin')->user()->school_id;
-        $subject->save();
-        return redirect('/admin/subject')->with('success','U shtua lenda');
+        if(Auth::guard('admin')->user()->hasPermissionTo('create-subject', 'admin')){
+            $this->validate($request,[
+                'Viti'=> 'required|numeric',
+                'Emri'=> 'required|string|min:2',
+            ]);
+            $subject = new Subject;
+            $subject->year = $request->input('Viti');
+            $subject->name = $request->input('Emri');
+            $subject->school_id = Auth::guard('admin')->user()->school_id;
+            $subject->save();
+            return redirect()->back()->with('success',__('messages.subject-add'));
+        }
+        else
+            return redirect(route('admin.home'))->with('error',__('messages.noauthorization'));
     }
 
     /**
@@ -64,10 +68,10 @@ class SubjectController extends Controller
     public function show($id)
     {
         $subject = Subject::find($id);
-        if(Auth::guard('admin'))
+        if(Auth::guard('admin')->user()->hasPermissionTo('view-subject', 'admin'))
             return view('admin.subject.show')->with('subject',$subject);
         else
-            return redirect('/home')->with('error','No access');
+            return redirect(route('admin.home'))->with('error',__('messages.noauthorization'));
     }
 
     /**
@@ -79,10 +83,10 @@ class SubjectController extends Controller
     public function edit($id)
     {
         $subject = Subject::find($id);
-        if(Auth::guard('admin'))
+        if(Auth::guard('admin')->user()->hasPermissionTo('edit-subject', 'admin'))
             return view('admin.subject.edit')->with('subject',$subject);
         else
-            return redirect('/home')->with('error','No access');
+            return redirect(route('admin.home'))->with('error',__('messages.noauthorization'));
     }
 
     /**
@@ -94,15 +98,20 @@ class SubjectController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request,[
-            'Viti'=> 'required|numeric',
-            'Emri'=> 'required|string|min:2',
-        ]);
-        $subject = Subject::find($id);
-        $subject->year = $request->input('Viti');
-        $subject->name = $request->input('Emri');
-        $subject->save();
-        return redirect('/admin/subject')->with('success','U ndryshua lenda');
+        if(Auth::guard('admin')->user()->hasPermissionTo('edit-subject', 'admin'))
+        {
+            $this->validate($request,[
+                'Viti'=> 'required|numeric',
+                'Emri'=> 'required|string|min:2',
+            ]);
+            $subject = Subject::find($id);
+            $subject->year = $request->input('Viti');
+            $subject->name = $request->input('Emri');
+            $subject->save();
+            return redirect()->back()->with('success',__('messages.subject-edit'));
+        }
+        else
+            return redirect(route('admin.home'))->with('error',__('messages.noauthorization'));
     }
 
     /**
@@ -113,10 +122,12 @@ class SubjectController extends Controller
      */
     public function destroy($id)
     {
-        $subject = Subject::find($id);
-        $subject->admins()->detach();
-        $subject->schedules()->delete();
-        $subject->delete();
-        return redirect('/admin/subject')->with('success','U fshi lenda');
+        if(Auth::guard('admin')->user()->hasPermissionTo('delete-subject', 'admin')){
+            $subject = Subject::find($id);
+            $subject->delete();
+            return redirect()->back()->with('success',__('messages.subject-delete'));
+        }
+        else
+            return redirect(route('admin.home'))->with('error',__('messages.noauthorization'));
     }
 }
